[REFACT] HeaderType과 HeaderTypeConfig 분리 #NADA-52
