+++ src/Bar/containers/HeaderType.js
@@ -1,6 +1,5 @@
 /*HeaderType 상단바의 유형을 관리함. 
 유형에 대하여는 컴플루언스-개발-프론트엔드개발-상단하단바관련 문서를 참고해주세요.*/
-import { Overlap, Img, Textwrapper } from '../styles/Header';
 
 //상단 상태값에서 상단 요소의 존재여부 값
 export const HeaderType = [
@@ -85,49 +84,3 @@ export const HeaderType = [
     post: true,
   },
 ];
-
-//각 요소에 대한 구현
-export const HeaderTypeConfig = (
-  navigate,
-  status,
-  pageNameStatus,
-  position,
-) => {
-  const Config = {
-    Texts: <Textwrapper>{pageNameStatus}</Textwrapper>,
-    Back: (
-      <Overlap>
-        <Img onClick={() => navigate(-1)} className="Back" />
-      </Overlap>
-    ),
-    Menu: (
-      <Overlap>
-        <Img className="Menu" />
-      </Overlap>
-    ),
-    Bell: (
-      <Overlap>
-        <Img className="Bell" />
-      </Overlap>
-    ),
-    BellWMenu: (
-      <Overlap>
-        <Img className="BellWMenu" />
-      </Overlap>
-    ),
-    Logo: (
-      <Overlap>
-        <Img className="Logo" />
-      </Overlap>
-    ),
-    Post: (
-      <Overlap>
-        <button className="Post" type="submit" form={position}>
-          완료
-        </button>
-      </Overlap>
-    ),
-  };
-
-  return Config[status.status];
-};
